Try a generalized approach for preinstalled premium plugins
