+++ client/my-sites/plugins/constants.js
@@ -1,3 +1,8 @@
+import {
+	PRODUCT_JETPACK_SEARCH_MONTHLY,
+	WPCOM_FEATURES_INSTANT_SEARCH,
+} from '@automattic/calypso-products';
+
 export const PREINSTALLED_PLUGINS = [
 	'jetpack',
 	'akismet',
@@ -7,3 +12,10 @@ export const PREINSTALLED_PLUGINS = [
 	'layout-grid',
 	'page-optimize',
 ];
+
+export const PREINSTALLED_PREMIUM_PLUGINS = {
+	'jetpack-search': {
+		feature: WPCOM_FEATURES_INSTANT_SEARCH,
+		product: PRODUCT_JETPACK_SEARCH_MONTHLY,
+	},
+};
